Update workspace.xml and .gitignore configurations
This commit is related to updates in the .idea/workspace.xml configuration file and .gitignore. The workspace.xml has new options set in "ChangeListManager" and "FileTemplateManagerImpl" components, and also includes a new component "TypeScriptGeneratedFilesManager". This enables the tracking of changes in more elements and improves the project administration. Meanwhile, in the .gitignore file, .idea/sonarlint/ has been added to the list of ignored files which further improves the project's version control system by preventing unnecessary files from being tracked.

Signed-off-by: GIUSEPPE BIANCONI <gbianconi3@gmail.com>